+++ lib_sonar/src/main/java/com/antoco/lib_sonar/view/sonarview/SonarGlView.kt
@@ -72,7 +72,7 @@ class SonarGlView : GLSurfaceView,Runnable {
         }
     }
 
-    fun setRange(range : Int){
+    fun setRange(range : Float){
         if(SonarSpec.range != range){
             clear()
             SonarSpec.range = range
